Refactor MJML email models and views for better organization
Moves specific email templates and their models (e.g., `SampleCard`) into a dedicated `Samples` sub-folder and namespace. Also relocates shared MJML email partials into an `Emails/Shared` folder.

This separation improves the logical grouping of email components, distinguishing generic elements from specific template implementations.
