+++ Anlab.Mvc/Models/Email/MjmlEmailTemplateModels.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-
 namespace AnlabMvc.Models.Email
 {
     public abstract class MjmlEmailTemplateModelBase
@@ -9,21 +7,6 @@ namespace AnlabMvc.Models.Email
         public string ButtonText { get; set; }
         public string ButtonUrl { get; set; }
         public string LayoutWidth { get; set; }
-    }
-
-    public class SampleCardEmailModel : MjmlEmailTemplateModelBase
-    {
-        public string Header { get; set; }
-        public string Message { get; set; }
-        public string CardTitle { get; set; }
-        public string CardSummary { get; set; }
-        public IReadOnlyList<SampleCardEmailItem> Items { get; set; } = new List<SampleCardEmailItem>();
-    }
-
-    public class SampleCardEmailItem
-    {
-        public string Label { get; set; }
-        public string Value { get; set; }
     }
 
     public class MjmlEmailButtonModel
